requirements.txt
streamlit
pandas
smtplib  # Note: This is part of Python's standard library and doesn't need to be installed
PyPDF2
openpyxl
seaborn
matplotlib
python-docx
python-pptx
celery
redis
msal
requests
